requirements.txt
alabaster==0.7.12
anyascii==0.1.7
appdirs==1.4.4
arrow==0.17.0
asgiref==3.3.1
Babel==2.9.0
bcrypt==3.2.0
beautifulsoup4==4.8.2
binaryornot==0.4.4
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==8.0.1
cookiecutter==1.7.2
distlib==0.3.1
distro==1.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.6
django-cache-url==2.0.0
django-filter==2.4.0
django-heroku==0.3.1
django-modelcluster==5.1
django-redis==4.11.0
django-storages==1.8
django-taggit==1.3.0
django-toolbelt==0.0.1
django-treebeard==4.4
djangorestframework==3.12.2
docker==4.4.1
docker-compose==1.28.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
draftjs-exporter==2.1.7
elasticsearch==2.4.1
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.1
idna==2.10
imagesize==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
jinja2-time==0.2.0
l18n==2020.6.1
MarkupSafe==1.1.1
openpyxl==3.0.6
packaging==20.9
paramiko==2.7.2
Pillow==8.1.0
pluggy==0.13.1
poyo==0.5.0
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2021.1
radar==0.3
redis==3.5.3
requests==2.25.1
selenium==3.141.0
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2
Sphinx==3.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
text-unidecode==1.3
texttable==1.6.3
toml==0.10.2
tox==3.21.4
Unidecode==1.1.1
urllib3==1.26.3
uWSGI==2.0.19.1
virtualenv==20.4.2
wagtail==2.12
wagtailfontawesome==1.1.4
webencodings==0.5.1
websocket-client==0.57.0
whitenoise==5.2.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
